backend/requirements.txt
mcp[cli]
fastapi
uvicorn[standard]
requests
feedparser
python-dotenv
